+++ github/github.go
@@ -40,14 +40,14 @@ func (api *API) UserExists(username string) (bool, error) {
 	return true, nil
 }
 
-func (api *API) AddUserToTeam(username string, teamID int64) error {
+func (api *API) AddUserToTeam(username string, teamID int) error {
 	_, _, err := api.Organizations.AddTeamMembership(
 		context.Background(), teamID, username, nil,
 	)
 	return err
 }
 
-func (api *API) RemoveUserFromTeam(username string, teamID int64) error {
+func (api *API) RemoveUserFromTeam(username string, teamID int) error {
 	_, err := api.Organizations.RemoveTeamMembership(
 		context.Background(), teamID, username,
 	)
@@ -76,7 +76,7 @@ func (api *API) CreateTeam(name string) (*gh.Team, error) {
 	return t, err
 }
 
-func (api *API) RemoveTeam(id int64) error {
+func (api *API) RemoveTeam(id int) error {
 	_, err := api.Organizations.DeleteTeam(context.Background(), id)
 	return err
 }
